src/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
databases==0.9.0
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.4.0
numpy==2.3.2
packaging==25.0
pip==25.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uuid==1.30
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
pytest
httpx
pytest-mock